+++ cli/docs/changelog.md
@@ -1,8 +1,8 @@
 # TBD
 
 ### Breaking Changes
-* Upgrade to core to 1.56.0 and engine to 1.28.0
-  * Users will need to run `kurtosis engine restart`
+* Upgrade to core to 1.57.0 and engine to 1.29.0
+  * Users will need to restart their Kurtosis engine
   
 ### Changes
 * Add test for setting resource allocations when adding a service to an enclave to internal testsuite